liburing.h: Populate SQE for {s,g} etsockopt
The current io_uring_prep_cmd_sock() liburing function is not populating
the SQE will all the fields because the kernel implementation was not
ready.

With the integration of the kernel part[1], populate the SQE with the
missing fields (optlen, optval, level and optname). This enables
the usage of this function to prepare commands that executes setsockopt
and getsockopt operations.

[1] Link:  https://lore.kernel.org/all/20231016134750.1381153-1-leitao@debian.org/
Signed-off-by: Breno Leitao <leitao@debian.org>