tests/socket-getsetsock-cmd: New test for {g,s}etsockopt
Introduce a new test to exercise the io_uring getsockopt and setsockopt
commands.

On the liburing side, use the io_uring_prep_cmd_sock() function to
prepare the SQE.

The test executes the same operation using a regular systemcall, such as
getsockopt(2), and the io_uring getsockopt command and compare the
results. The same for the setsockopt counterpart.

Signed-off-by: Breno Leitao <leitao@debian.org>